+++ packages/gatsby-source-wordpress/src/normalize.js
@@ -165,9 +165,16 @@ exports.liftRenderedField = entities =>
   })
 
 // Exclude entities of unknown shape
-// Assume all entities contain a wordpress_id, except for whitelisted type wp_settings
+// Assume all entities contain a wordpress_id,
+// except for whitelisted type wp_settings and the site_metadata
 exports.excludeUnknownEntities = entities =>
-  entities.filter(e => e.wordpress_id || e.__type === `wordpress__wp_settings`) // Excluding entities without ID, or WP Settings
+  entities.filter(
+    e =>
+      e.wordpress_id ||
+      e.__type === `wordpress__wp_settings` ||
+      e.__type === `wordpress__site_metadata`
+  )
+// Excluding entities without ID, or WP Settings
 
 // Create node ID from known entities
 // excludeUnknownEntities whitelisted types don't contain a wordpress_id
@@ -519,10 +526,10 @@ const prepareACFChildNodes = (
   _.each(obj, (value, key) => {
     if (_.isArray(value) && value[0] && value[0].acf_fc_layout) {
       obj[`${key}___NODE`] = value.map(
-        v =>
+        (v, indexItem) =>
           prepareACFChildNodes(
             v,
-            entityId,
+            `${entityId}_${indexItem}`,
             topLevelIndex,
             type + key,
             children,