+++ packages/gatsby/src/schema/index.js
@@ -7,8 +7,10 @@ const buildNodeTypes = require(`./build-node-types`)
 const buildNodeConnections = require(`./build-node-connections`)
 const { store } = require(`../redux`)
 const invariant = require(`invariant`)
+const { clearUnionTypes } = require(`./infer-graphql-type`)
 
 module.exports = async ({ parentSpan }) => {
+  clearUnionTypes()
   const typesGQL = await buildNodeTypes({ parentSpan })
   const connections = buildNodeConnections(_.values(typesGQL))