+++ www/src/components/package-readme.js
@@ -24,10 +24,10 @@ class PackageReadMe extends React.Component {
         <Helmet>
           <title>{packageName}</title>
           <meta name="description" content={excerpt} />
-          <meta name="og:description" content={excerpt} />
+          <meta property="og:description" content={excerpt} />
           <meta name="twitter:description" content={excerpt} />
-          <meta name="og:title" content={packageName} />
-          <meta name="og:type" content="article" />
+          <meta property="og:title" content={packageName} />
+          <meta property="og:type" content="article" />
           <meta name="twitter.label1" content="Reading time" />
           <meta name="twitter:data1" content={`${timeToRead} min read`} />
         </Helmet>
@@ -50,11 +50,11 @@ class PackageReadMe extends React.Component {
           >
             <GithubIcon style={{ verticalAlign: `text-top` }} />
           </a>
-          {/* {githubUrl && (
+          {githubUrl && (
             <Link to={`/starters?d=${packageName}`}>
               See starters that use this
             </Link>
-          )} */}
+          )}
         </div>
 
         <div